+++ CHANGELOG.md
@@ -6,9 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+## [1.5.0] - 2022-08-02
+### Added
 - BREAKING: Add support for fine grained include/exclude definitions AKA white/gray/blacklisting (#5). Please note that the old `ignore_disks:` section is no longer available. You have to move those configuration values into the `disks.exclude:` section
 - Disks can be included and excluded with regular expressions (#6) 
 - Files can now be sorted by `born_at`, `modified_at`, `archived_at` and `interpolation`. Configuration parameters `defaults.sort` and `files.*.sort` are  activated.
+- Disks can be refreshed by hitting `Ctrl+R` or just `r` in the console
+
+### Fixed
+- `--debug` had no effect when log setting in `config.yaml` was missing
 
 ## [1.4.0] - 2022-07-19
 ### Added